refactor(theme): 重构暗黑模式切换组件并优化动画效果
重构暗黑模式切换组件，将 `Dark_mode_switch.vue` 重命名为 `switch.vue`，并优化了动画效果和代码结构。同时，调整了 CSS 动画的定义，使其更加简洁和一致。
