+++ .vitepress/theme/index.ts
@@ -1,7 +1,8 @@
 import { h } from 'vue'
 import type { Theme } from 'vitepress'
-import MyLayout from './components/Dark_mode_switch.vue'
 import DefaultTheme from 'vitepress/theme'
+import MyLayout from './components/switch.vue'
+
 import './css/background.css'
 import './css/text.css'
 import './css/brand.css'
